refactor(complete): move completion_verification into gptme.util
Helpers belong in gptme.util, not the package top-level.
