+++ gptme/tools/complete.py
@@ -13,7 +13,11 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any
 from xml.sax.saxutils import escape as xml_escape
 
-from ..completion_verification import (
+from ..hooks import HookType, StopPropagation
+from ..hooks.confirm import ConfirmAction, get_confirmation
+from ..message import Message
+from ..sandbox import SandboxConfig, build_env, wrap_shell_cmd
+from ..util.completion_verification import (
     ApprovedBind,
     VerificationCommand,
     approved_execution_argv,
@@ -23,10 +27,6 @@ from ..completion_verification import (
     restore_approved_snapshots,
     uses_approved_snapshot,
 )
-from ..hooks import HookType, StopPropagation
-from ..hooks.confirm import ConfirmAction, get_confirmation
-from ..message import Message
-from ..sandbox import SandboxConfig, build_env, wrap_shell_cmd
 from .base import ToolSpec, ToolUse
 from .shell_validation import is_denylisted
 from .todo import get_incomplete_todos_summary, has_incomplete_todos
